+++ ferrite-core/src/ui/input.rs
@@ -29,7 +29,7 @@ pub fn handle_input(
     }
 }
 
-fn handle_zoom(ui: &Ui, zoom_handler: &mut ZoomHandler, scroll_delta: f32) {
+fn handle_zoom(ui: &Ui, zoom_handler: &mut ZoomHandler, scroll_delta: f64) {
     if let Some(mouse_pos) = ui.input(|i| i.pointer.hover_pos()) {
         let zoom_step = if scroll_delta > 0.0 { 1.1 } else { 0.9 };
         let new_zoom = (zoom_handler.zoom_level() * zoom_step).clamp(0.1, 10.0);
